pyproject.toml
[project]
name = "clockcaptcha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "diffusers",
    "transformers",
    "accelerate",
    "safetensors",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "jinja2",
    "pyjwt",
    "python-jose[cryptography]",
    "opencv-python",
    "pillow",
    "numpy",
    "python-dotenv",
]
